+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 - [#2048](https://gitlab.com/meltano/meltano/-/issues/2048) Add `docker` and `gitlab-ci` file bundles to allow instant containerization
 - [#2068](https://gitlab.com/meltano/meltano/-/issues/2068) Add interface to schedule new pipelines to Pipelines page
+- [#2081](https://gitlab.com/meltano/meltano/-/issues/2081) Add `url` config option to `target-postgres`
 
 ### Changes